shopping_info
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@3310 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/service/NnAdManager.java
@@ -5,6 +5,7 @@ import java.util.logging.Logger;
 
 import com.nncloudtv.dao.NnAdDao;
 import com.nncloudtv.model.NnAd;
+import com.nncloudtv.model.NnEpisode;
 
 public class NnAdManager {
     
@@ -12,24 +13,57 @@ public class NnAdManager {
     
     private NnAdDao dao = new NnAdDao();
     
-    public NnAd findByProgramId(long programId) {
-        return dao.findByProgramId(programId);
+    public NnAd findByEpisode(NnEpisode episode) {
+        
+        if (episode == null) {
+            return null;
+        }
+        
+        long adId = episode.getAdId();
+        
+        if (adId == 0) {
+            return null;
+        }
+        
+        return dao.findById(adId);
+        
     }
     
     public void delete(NnAd nnad) {
+        
+        if (nnad == null) {
+            return;
+        }
+        
         dao.delete(nnad);
     }
     
-    public NnAd save(NnAd nnad) {
+    public NnAd save(NnAd nnad, NnEpisode episode) {
+        
+        if (nnad == null || episode == null) {
+            return null;
+        }
         
         Date now = new Date();
         
         if (nnad.getCreateDate() == null) {
             nnad.setCreateDate(now);
         }
-        nnad.setUpdateDate(now);
         
-        return dao.save(nnad);
+        nnad.setUpdateDate(now);
+        nnad = dao.save(nnad);
+        
+        NnAd origAd = findByEpisode(episode);
+        if (origAd != null && origAd.getId() != nnad.getId()) {
+            
+            delete(origAd);
+        }
+        
+        NnEpisodeManager episodeMngr = new NnEpisodeManager();
+        episode.setAdId(nnad.getId());
+        episodeMngr.save(episode);
+        
+        return nnad;
     }
     
 }
